Move c-abi testing timeout utils
