+++ libvcx/src/api/return_types_u32.rs
@@ -15,7 +15,7 @@ use crate::libindy::utils::callback_u32 as callback;
 use crate::libindy::utils::error_codes::map_indy_error;
 use crate::libindy::utils::next_command_handle;
 use crate::utils::error;
-use crate::utils::timeout::TimeoutUtils;
+use crate::abi_utils::timeout::TimeoutUtils;
 
 fn log_error<T: Display>(e: T) {
     warn!("Unable to send through libindy callback in vcx: {}", e);
