+++ libvcx/src/api/utils.rs
@@ -549,7 +549,7 @@ mod tests {
     use crate::api::return_types_u32;
     use crate::utils::constants;
     use crate::utils::devsetup::*;
-    use crate::utils::timeout::TimeoutUtils;
+    use crate::abi_utils::timeout::TimeoutUtils;
 
     use super::*;
     use crate::utils::provision::AgentProvisionConfig;
