+++ libvcx/src/utils/mod.rs
@@ -36,7 +36,6 @@ macro_rules! map (
 
 pub mod error;
 pub mod constants;
-pub mod timeout;
 pub mod openssl;
 pub mod json;
 pub mod runtime;
